Fix warnings

+++ buildSrc/src/main/java/com/kizitonwose/calendar/buildsrc/Build.kt
@@ -7,7 +7,7 @@ import org.gradle.jvm.toolchain.JavaLanguageVersion
 
 object Config {
     val compatibleJavaVersion = JavaVersion.VERSION_17
-    val compatibleJavaLanguageVersion = JavaLanguageVersion.of(compatibleJavaVersion.majorVersion.toInt())
+    val compatibleJavaLanguageVersion = JavaLanguageVersion.of(compatibleJavaVersion.majorVersion)
 }
 
 object Version {
